add onError to ControllerRouteOptions type
