+++ src/controllers/DELETE.ts
@@ -17,7 +17,7 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import type { ControllerRouteWithBodyOptions, Nilable } from '../types';
-import { createHttpMethodDecorator } from './utils';
+import { createHttpMethodDecorator } from './factories';
 
 /**
  * Add a controller method to handle a DELETE request.
