+++ src/controllers/utils.ts
@@ -14,93 +14,7 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import path from 'path';
-import { INIT_CONTROLLER_METHOD_ACTIONS } from '../constants';
-import type { ControllerRouteOptionsValue, ControllerRouteWithBodyOptions, HttpMethod, HttpMiddleware, HttpRequestHandler, HttpRequestPath, Nilable } from '../types';
-import type { InitControllerMethodAction } from '../types/internal';
-import { asAsync, isNil } from '../utils';
-import { params } from '../validators';
-
-interface ICreateControllerMethodRequestHandlerOptions {
-    method: Function;
-}
-
-export interface ICreateHttpMethodDecoratorOptions {
-    decoratorOptions: Nilable<ControllerRouteOptionsValue<ControllerRouteWithBodyOptions>>;
-    name: HttpMethod;
-}
-
-interface ICreateInitControllerMethodActionOptions {
-    controllerMethod: string;
-    controllerRouterPath: Nilable<string>;
-    httpMethod: HttpMethod;
-    middlewares: HttpMiddleware[];
-}
-
-function createControllerMethodRequestHandler({ method }: ICreateControllerMethodRequestHandlerOptions): HttpRequestHandler {
-    method = asAsync<HttpRequestHandler>(method);
-
-    return (request, response) => method(request, response);
-}
-
-export function createHttpMethodDecorator(options: ICreateHttpMethodDecoratorOptions): MethodDecorator {
-    return function (target, methodName, descriptor) {
-        const method = getMethodOrThrow(descriptor);
-
-        const decoratorOptions: Nilable<ControllerRouteWithBodyOptions> =
-            typeof options.decoratorOptions === 'string' ? {
-                path: options.decoratorOptions
-            } : options.decoratorOptions;
-
-        let initActions: Nilable<InitControllerMethodAction[]> = (method as any)[INIT_CONTROLLER_METHOD_ACTIONS];
-        if (!initActions) {
-            (method as any)[INIT_CONTROLLER_METHOD_ACTIONS] = initActions = [];
-        }
-
-        const middlewares = (
-            Array.isArray(decoratorOptions?.use) ?
-                decoratorOptions!.use :
-                [decoratorOptions?.use]
-        ).filter(mw => !isNil(mw)) as HttpMiddleware[];
-
-        initActions.push(createInitControllerMethodAction({
-            controllerMethod: String(methodName).trim(),
-            controllerRouterPath: decoratorOptions?.path,
-            httpMethod: options.name,
-            middlewares
-        }));
-    };
-}
-
-function createInitControllerMethodAction({ controllerMethod, controllerRouterPath, httpMethod, middlewares }: ICreateInitControllerMethodActionOptions): InitControllerMethodAction {
-    return ({ relativeFilePath, method, server }) => {
-        const dir = path.dirname(relativeFilePath);
-        const fileName = path.basename(relativeFilePath, path.extname(relativeFilePath));
-
-        let routerPath: HttpRequestPath = dir;
-        if (fileName !== 'index') {
-            routerPath += `/${fileName}`;
-        }
-
-        if (controllerRouterPath?.length) {
-            routerPath += normalizeRouterPath(controllerRouterPath);
-        } else {
-            if (controllerMethod.length && controllerMethod !== 'index') {
-                routerPath += `/${controllerMethod}`;
-            }
-        }
-
-        routerPath = normalizeRouterPath(routerPath);
-        routerPath = routerPath.split('/@').join('/:');
-
-        if (routerPath.includes('/:')) {
-            routerPath = params(routerPath);
-        }
-
-        (server as any)[httpMethod](routerPath, middlewares, createControllerMethodRequestHandler({
-            method
-        }));
-    };
-}
+import type { Nilable } from '../types';
 
 export function getMethodOrThrow<T extends Function = Function>(descriptor: PropertyDescriptor): T {
     const method: any = descriptor?.value;
